requirements.txt
websockets
pyperclip
Pillow